+++ src/eventserver/socketserver.cpp
@@ -366,7 +366,7 @@ int SocketServer::ReadLoop(){
         }
     }
 EXIT:
-    return ret;
+    return (int)ret;
 }
